Lesson 69 (commit 10) add change in fields price

+++ src/main/java/com/example/java_group_11_online_store_ayday_mirbekkyzy/DTO/BasketDTO.java
@@ -25,6 +25,6 @@ public class BasketDTO {
     private Integer id;
     private Products products_id;
     private User customers_id;
-    private Float price;
+    private Integer price;
     private Integer quantity;
 }
