+++ src/main/resources/db/migration/V2022.05.22.14.36__add_table_baskets.sql
@@ -4,7 +4,7 @@ create table `baskets`(
                          `customers_id` int(10),
                          `products_id` int(10),
                          `quantity` int(10),
-                         `price` float(10),
+                         `price` int (10),
                          primary key (`id`),
                          constraint `fk_customer_busket` foreign key (`customers_id`) references `user_table` (`id`),
                          constraint `fk_products_basket` foreign key (`products_id`) references `products` (`id`)
